# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Top deal shop dự án tốt nghiệp.",
        "description": "https://apitopdeal.shop/",
        "contact": {
            "name": "Tên người liên hệ",
            "email": "email@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Lấy thông tin người dùng",
                "description": "Lấy thông tin chi tiết của người dùng, bao gồm cả địa chỉ liên kết. Chỉ admin hoặc chính người dùng có thể xem thông tin của họ.",
                "operationId": "getUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của người dùng",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lấy thông tin thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lấy thông tin thành công"
                                        },
                                        "users": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Người dùng không có quyền truy cập",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Bạn không có quyền xem thông tin của người dùng khác."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Chưa đăng nhập",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Bạn cần đăng nhập để xem thông tin."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy người dùng",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Không tồn tại người dùng này."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Lỗi máy chủ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Đã xảy ra lỗi không mong muốn."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/profile": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Lấy thông tin hồ sơ người dùng hiện tại",
                "operationId": "c65ebd108ed135fde20b650a5a5b676e",
                "responses": {
                    "200": {
                        "description": "Thông tin hồ sơ người dùng hiện tại được lấy thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Người dùng không tìm thấy",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "User not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Có lỗi xảy ra",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/auth/profile/update/{id}": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Cập nhật thông tin người dùng",
                "operationId": "38045b28c1494391ea90eca9c9eae139",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của người dùng",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProfileRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật thông tin thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Cập nhật thông tin thành công."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không thể thay đổi admin",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Không thể thay đổi admin"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Có lỗi xảy ra",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Cập nhật thông tin thất bại"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/auth/addresses/destroy/{id}": {
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Xóa địa chỉ của người dùng",
                "operationId": "4ef2adb52525298a3361f1fcf1534748",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của địa chỉ cần xóa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Xóa địa chỉ thành công"
                    },
                    "403": {
                        "description": "Không thể xóa địa chỉ mặc định"
                    },
                    "404": {
                        "description": "Người dùng không tồn tại hoặc địa chỉ không tồn tại"
                    },
                    "500": {
                        "description": "Có lỗi xảy ra"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Nguyễn Văn A"
                    },
                    "email": {
                        "type": "string",
                        "example": "email@example.com"
                    },
                    "avatar": {
                        "type": "string",
                        "example": "path/to/avatar.png"
                    },
                    "addresses": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Address"
                        }
                    }
                },
                "type": "object"
            },
            "UpdateProfileRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Nguyễn Văn A"
                    },
                    "email": {
                        "type": "string",
                        "example": "email@example.com"
                    },
                    "avatar": {
                        "type": "string",
                        "example": "path/to/avatar.png"
                    },
                    "addresses": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Address"
                        }
                    }
                },
                "type": "object"
            },
            "Address": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "address_name": {
                        "type": "string",
                        "example": "Số 1, Đường ABC"
                    },
                    "city": {
                        "type": "string",
                        "example": "Hà Nội"
                    },
                    "district": {
                        "type": "string",
                        "example": "Quận Hoàn Kiếm"
                    },
                    "ward": {
                        "type": "string",
                        "example": "Phường Tràng Tiền"
                    },
                    "detail_address": {
                        "type": "string",
                        "example": "Gần Hồ Gươm"
                    },
                    "is_default": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "description": "Nhập token JWT trong định dạng 'Bearer {token}'",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}